TRIPS-46 - Implement 'travel map' component
* Add reducers scaffolding
