+++ client/app.jsx
@@ -1,10 +1,16 @@
 import React from 'react';
 import { render } from 'react-dom';
+import { createStore, applyMiddleware } from 'redux';
+import { Provider } from 'react-redux';
+import createSagaMiddleware from 'redux-saga';
 import Trips from './modules/Trips';
 import TravelMap from './modules/TravelMap';
+import travelMapReducer from './modules/TravelMap/reducers';
 
 const tripsContainer = document.querySelector('.js-trips');
 const mapContainer = document.querySelector('.js-travel-map');
+
+const sagaMiddleware = createSagaMiddleware();
 
 if (tripsContainer) {
   render(
@@ -14,10 +20,15 @@ if (tripsContainer) {
 }
 
 if (mapContainer) {
+  const store = createStore(
+      travelMapReducer,
+      applyMiddleware(sagaMiddleware)
+  );
+
   render(
-    <TravelMap
-      countries={['GB', 'DE', 'US']}
-    />,
+    <Provider store={store}>
+      <TravelMap />
+    </Provider>,
     mapContainer
   );
 }
