+++ client/modules/TravelMap/reducers.js
@@ -1,0 +1,10 @@
+const defaultState = {
+  countries: [],
+};
+
+export default function reducer(state = defaultState, action) {
+  switch (action.type) {
+    default:
+      return state;
+  }
+}
