+++ package.json
@@ -49,6 +49,9 @@
     "precss": "^1.4.0",
     "react": "^15.4.2",
     "react-dom": "^15.4.2",
+    "react-redux": "^5.0.7",
+    "redux": "^3.7.2",
+    "redux-saga": "^0.16.0",
     "sass-loader": "^4.0.1",
     "selenium-webdriver": "^3.5.0",
     "style-loader": "^0.13.1",
